Struct simple_endian::LittleEndian

#[repr(transparent)]
pub struct LittleEndian<T: SpecificEndian<T>> { /* private fields */ }

A little-endian representation of type T that implements SpecificEndian<T>. Data stored in the struct must be converted to little-endian using ::from() or .into().

Implementations

impl<T> LittleEndian<T>where T: SpecificEndian<T>,

pub fn to_bits(&self) -> T

Returns the raw data stored in the struct.

pub fn from_bits(v: T) -> Self

Imports the data raw into a LittleEndian struct.

pub fn to_native(&self) -> T

Converts the data to the same type T in host-native endian.